Changed ExposedVideoData to use void* under Windows, fixes problems on 64bit systems. Some other 64bit warnings silenced. All changes provided by Kosuha.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@1313 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/dmfsupport.h
@@ -168,7 +168,7 @@ public:
 		{
 			if (*p == '\n')
 			{
-				core::stringc str(start, p - start - 1);
+				core::stringc str(start, (u32)(p - start - 1));
 				str.trim();
 				Add(str);
 				start = p+1;
@@ -179,7 +179,7 @@ public:
 
 		if (p - start > 1)
 		{
-			core::stringc str(start, p - start - 1);
+			core::stringc str(start, (u32)(p - start - 1));
 			str.trim();
 			Add(str);
 		}
